fix: added browser API for FE functions
