+++ api/fe.js
@@ -1,0 +1,11 @@
+const { generateScript } = require('../forward_engineering/api/generateScript');
+const { generateViewScript } = require('../forward_engineering/api/generateViewScript');
+const { generateContainerScript } = require('../forward_engineering/api/generateContainerScript');
+const { isDropInStatements } = require('../forward_engineering/api/isDropInStatements');
+
+module.exports = {
+	generateScript,
+	generateViewScript,
+	generateContainerScript,
+	isDropInStatements,
+};
